server-fastapi/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic[email]==2.10.3
bcrypt==4.2.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.18
pyjwt==2.10.1
scikit-learn==1.5.2
sentence-transformers==2.2.2
